Move tests from StateBinderTest to ScopesTest
